gas-station/generator/requirements.txt
Faker==33.1.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
tzdata==2025.1
